% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk_on_simple_portfolio.R
\name{aggregate_simple_portfolio}
\alias{aggregate_simple_portfolio}
\title{Aggregate simple portfolio input rows}
\usage{
aggregate_simple_portfolio(portfolio_data)
}
\arguments{
\item{portfolio_data}{A simple portfolio dataframe.}
}
\value{
Portfolio aggregated at company/term/lgd level.
}
\description{
Aggregate simple portfolio input rows
}
