% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk_on_simple_portfolio.R
\name{aggregate_simple_portfolio_results}
\alias{aggregate_simple_portfolio_results}
\title{Aggregate simple portfolio tech details back to input-row shape}
\usage{
aggregate_simple_portfolio_results(portfolio_results_tech_detail)
}
\arguments{
\item{portfolio_results_tech_detail}{Detailed output from \code{run_trisk_on_simple_portfolio()}.}
}
\value{
Portfolio output with one row per input portfolio row.
}
\description{
Aggregate simple portfolio tech details back to input-row shape
}
