% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk_on_simple_portfolio.R
\name{aggregate_trisk_outputs_simple}
\alias{aggregate_trisk_outputs_simple}
\title{Aggregate TRISK outputs for simple portfolio analysis}
\usage{
aggregate_trisk_outputs_simple(npv_results, pd_results)
}
\arguments{
\item{npv_results}{NPV results from \code{trisk.model::run_trisk_model()}.}

\item{pd_results}{PD results from \code{trisk.model::run_trisk_model()}.}
}
\value{
TRISK outputs aggregated without asset/country dimensions.
}
\description{
Aggregate TRISK outputs for simple portfolio analysis
}
