% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk_on_simple_portfolio.R
\name{compute_simple_portfolio_metrics}
\alias{compute_simple_portfolio_metrics}
\title{Compute value-change and expected-loss metrics on exposure shares}
\usage{
compute_simple_portfolio_metrics(analysis_data)
}
\arguments{
\item{analysis_data}{Analysis dataset with exposure shares.}
}
\value{
Analysis dataset with additional metrics columns.
}
\description{
Compute value-change and expected-loss metrics on exposure shares
}
