% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tech_mix.R
\name{draw_trisk_tech_mix}
\alias{draw_trisk_tech_mix}
\title{Draw the technology-mix plot}
\usage{
draw_trisk_tech_mix(mix, bases = c("\% of NPV value", "\% of exposure (EAD)"))
}
\arguments{
\item{mix}{Output of [prepare_for_trisk_tech_mix()].}

\item{bases}{Bases to draw (column names present in `mix`).}
}
\value{
A ggplot object.
}
\description{
Draw the technology-mix plot
}
\keyword{internal}
