% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_params.R
\name{get_available_parameters}
\alias{get_available_parameters}
\title{Get the possible parameters values for Trisk from the scenario data}
\usage{
get_available_parameters(scenarios)
}
\arguments{
\item{scenarios}{scenarios dataframe}
}
\value{
Available scenarios parameters
}
\description{
Get the possible parameters values for Trisk from the scenario data
}
