% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sensitivity_analysis.R
\name{get_filtered_assets_data}
\alias{get_filtered_assets_data}
\title{Get filtered assets data based on provided criteria}
\usage{
get_filtered_assets_data(
  input_path,
  country_iso2 = NULL,
  sector = NULL,
  technology = NULL,
  company_name = NULL
)
}
\arguments{
\item{input_path}{The path to the input data directory containing the necessary files for the TRISK model.}

\item{country_iso2}{A character vector of ISO2 country codes to filter the assets. Defaults to NULL.}

\item{sector}{A character vector of sectors to filter the assets. Defaults to NULL.}

\item{technology}{A character vector of technologies to filter the assets. Defaults to NULL.}

\item{company_name}{A character vector of company names to filter the assets. Defaults to NULL.}
}
\value{
A list containing the filtered assets data and other input data required for the TRISK model.
}
\description{
This function loads and filters the assets data based on criteria such as country, sector, technology, and company name.
}
