% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_trisk_result.R
\name{get_latest_timestamped_folder}
\alias{get_latest_timestamped_folder}
\title{Get the latest timestamped folder in a directory}
\usage{
get_latest_timestamped_folder(path)
}
\arguments{
\item{path}{The directory path containing timestamped folders.}
}
\value{
The relative path of the most recent timestamped folder. If no valid directories are found, a message is returned.
}
\description{
This function finds the most recent folder in a given directory where the folder names follow a specific
timestamp format (YYYYMMDD_HHMMSS).
}
