% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_trisk_result.R
\name{get_latest_trisk_result}
\alias{get_latest_trisk_result}
\title{Get the latest TRISK result of a specified type}
\usage{
get_latest_trisk_result(trisk_output_path, result_type)
}
\arguments{
\item{trisk_output_path}{The path to the TRISK output directory containing timestamped folders.}

\item{result_type}{The type of result to retrieve. Valid options are "npv", "pd", "trajectories", and "params".}
}
\value{
A tibble containing the requested TRISK result data. If the specified file does not exist, a warning is issued, and NULL is returned.
}
\description{
This function retrieves the latest TRISK result based on the specified result type.
It identifies the most recent output directory using a timestamped folder structure, and then loads
the corresponding result file.
}
