% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_trisk_result.R
\name{get_single_trisk_result}
\alias{get_single_trisk_result}
\title{Get a specific TRISK result from a given sub-directory}
\usage{
get_single_trisk_result(sub_dir, result_type)
}
\arguments{
\item{sub_dir}{The sub-directory path containing the TRISK result files.}

\item{result_type}{The type of result to retrieve. Valid options are "trajectories", "pd", "npv", and "params".}
}
\value{
A tibble containing the requested TRISK result data. If the file does not exist, a warning is issued, and NULL is returned.
}
\description{
This function retrieves a specific TRISK result file from a given sub-directory based on the provided result type.
}
