% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk_on_simple_portfolio.R
\name{join_simple_portfolio_to_trisk}
\alias{join_simple_portfolio_to_trisk}
\title{Join simple portfolio to aggregated TRISK outputs}
\usage{
join_simple_portfolio_to_trisk(portfolio_data, trisk_agg)
}
\arguments{
\item{portfolio_data}{Aggregated simple portfolio.}

\item{trisk_agg}{Aggregated TRISK outputs.}
}
\value{
Joined dataframe at company/term/lgd/sector/technology level.
}
\description{
Join simple portfolio to aggregated TRISK outputs
}
