% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi_trajectories.R
\name{prepare_for_trisk_line_plot}
\alias{prepare_for_trisk_line_plot}
\title{Prepare Data for Risk Trajectory Line Plot}
\usage{
prepare_for_trisk_line_plot(trajectories_data, facet_var, linecolor)
}
\arguments{
\item{trajectories_data}{Dataset containing trajectory information across different scenarios.}

\item{facet_var}{Variable for faceting plots by business units.}

\item{linecolor}{Variable for coloring lines by sector.}
}
\value{
A dataframe ready for plotting, with production percentages for "production_shock_scenario" for visualization, excluding the last year of each "run_id" group.
}
\description{
Prepares trajectory data for line plot visualization, calculating production percentages of the maximum value for "production_shock_scenario" only,
and removing the last year of each "run_id" group.
}
