% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk_excel.R
\name{run_trisk_excel}
\alias{run_trisk_excel}
\title{Run TRISK model and save results to Excel}
\usage{
run_trisk_excel(input_path, output_path, ...)
}
\arguments{
\item{input_path}{Character. Path to the input file used by the TRISK model.}

\item{output_path}{Character. Path to the output folder where results will be saved.}

\item{...}{Additional arguments passed to `trisk.model::run_trisk`.}
}
\value{
None. The function saves the processed results as Excel or CSV files in the output directory and prints a completion message.
}
\description{
This function runs the TRISK model, processes the resulting CSV files, and converts them to Excel format if they contain fewer rows than Excel's row limit. If a file exceeds Excel's row limit, it is saved as a CSV with a new name.
}
