% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{run_trisk_model}
\alias{run_trisk_model}
\title{run_trisk_model}
\value{
See \code{trisk.model::run_trisk_model}
}
\description{
Re-exported function from trisk.model package. Please see \code{trisk.model::run_trisk_model} for full documentation.
}
\details{
Run TRISK Model
}
\seealso{
\code{\link[trisk.model]{run_trisk_model}}
}
