% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sensitivity_analysis.R
\name{run_trisk_sa}
\alias{run_trisk_sa}
\title{Run TRISK sensitivity analysis on multiple scenarios}
\usage{
run_trisk_sa(
  assets_data,
  scenarios_data,
  financial_data,
  carbon_data,
  run_params,
  ...
)
}
\arguments{
\item{assets_data}{Data frame containing asset information.}

\item{scenarios_data}{Data frame containing scenario information.}

\item{financial_data}{Data frame containing financial information.}

\item{carbon_data}{Data frame containing carbon price information.}

\item{run_params}{A list of parameter sets where each set contains the required parameters for a single TRISK model run.
Each parameter set must include `scenario_geography`, `baseline_scenario`, `target_scenario`, and `shock_year`. Find their definition and other Trisk parameters at \code{\link[trisk.model]{run_trisk_model}}}

\item{...}{Additional arguments passed to \code{\link{get_filtered_assets_data}} (`country_iso2`, `sector`, `technology`, and `company_name`).}
}
\value{
A list of tibbles containing the combined results for all runs. The list includes tibbles for NPV results (`npv`),
        PD results (`pd`), company trajectories (`trajectories`), and model parameters (`params`).
}
\description{
This function performs a sensitivity analysis by running the TRISK model on multiple scenarios.
It takes a list of parameter sets and runs the TRISK model for each set, returning a comprehensive
set of results that includes net present value (NPV), probability of default (PD), company trajectories, and model parameters.
}
