% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trisk_single_filtered.R
\name{run_trisk_single_filtered}
\alias{run_trisk_single_filtered}
\title{Run a single filtered TRISK model scenario}
\usage{
run_trisk_single_filtered(
  input_path,
  run_params,
  country_iso2 = NULL,
  sector = NULL,
  technology = NULL,
  company_name = NULL
)
}
\arguments{
\item{input_path}{The path to the input data directory containing the necessary files for the TRISK model.}

\item{run_params}{A list of parameters required for running the TRISK model, including `scenario_geography`, 
`baseline_scenario`, `target_scenario`, and `shock_year`.}

\item{country_iso2}{A character vector of ISO2 country codes to filter the assets. Defaults to NULL.}

\item{sector}{A character vector of sectors to filter the assets. Defaults to NULL.}

\item{technology}{A character vector of technologies to filter the assets. Defaults to NULL.}

\item{company_name}{A character vector of company names to filter the assets. Defaults to NULL.}
}
\value{
A list of tibbles containing the results of the TRISK model run. The list includes tibbles for NPV results (`npv`), 
        PD results (`pd`), company trajectories (`trajectories`), and model parameters (`params`).
}
\description{
This function runs the TRISK model for a single set of parameters, with the ability to filter assets 
based on country, sector, technology, and company name. The function returns a set of results including 
net present value (NPV), probability of default (PD), company trajectories, and model parameters.
}
