% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{TRISK_SECTOR_PALETTE}
\alias{TRISK_SECTOR_PALETTE}
\title{TRISK sector colour system}
\format{
Named character vector of hex colours (one base colour per sector).
}
\usage{
TRISK_SECTOR_PALETTE
}
\description{
A base hue per climate-relevant sector. Technologies within a sector are
rendered as shades of that base (darkest = most carbon-intensive) by
[trisk_sector_shades()], used in [pipeline_trisk_tech_mix()].
}
\keyword{datasets}
