% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot_data.R
\name{aggregate_facts_trisk}
\alias{aggregate_facts_trisk}
\title{Title}
\usage{
aggregate_facts_trisk(analysis_data, group_cols)
}
\arguments{
\item{analysis_data}{analysis_data}

\item{group_cols}{group_cols}
}
\description{
TODO FIND CLOSEST COMPANY IF group_cols=NULL
}
