% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trisk_inputs.R
\name{download_trisk_inputs}
\alias{download_trisk_inputs}
\title{Download TRISK input data files from a specified endpoint}
\usage{
download_trisk_inputs(local_save_folder)
}
\arguments{
\item{endpoint_url}{The base URL of the endpoint from which the data files will be downloaded.}
}
\value{
No return value. The function downloads files and saves them to the specified local folder.
}
\description{
This function downloads TRISK input data files from a specified endpoint URL and saves them to a local folder.
It ensures that the local folder exists before downloading the files. If the folder does not exist, it is created.
Env variables containing endpoints are defined in imports.R
}
