% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sensitivity_analysis.R
\name{get_filtered_assets_data}
\alias{get_filtered_assets_data}
\title{Filter assets based on provided criteria}
\usage{
get_filtered_assets_data(
  assets_data,
  country_iso2 = NULL,
  sector = NULL,
  technology = NULL,
  company_name = NULL
)
}
\arguments{
\item{assets_data}{The assets data to be filtered.}

\item{country_iso2}{A character vector of ISO2 country codes to filter the assets. Defaults to NULL.}

\item{sector}{A character vector of sectors to filter the assets. Defaults to NULL.}

\item{technology}{A character vector of technologies to filter the assets. Defaults to NULL.}

\item{company_name}{A character vector of company names to filter the assets. Defaults to NULL.}
}
\value{
A tibble containing the filtered assets data.
}
\description{
This function filters the assets data based on criteria such as country, sector, technology, and company name.
}
